+++ src/api/endpoints/orders/get-user-top-bids/v1.ts
@@ -104,6 +104,7 @@ export const getUserTopBidsV1Options: RouteOptions = {
             tokenId: Joi.string(),
             name: Joi.string().allow(null, ""),
             image: Joi.string().allow(null, ""),
+            floorAskPrice: Joi.number().unsafe().allow(null),
             collection: Joi.object({
               id: Joi.string().allow(null),
               name: Joi.string().allow(null, ""),
@@ -239,14 +240,15 @@ export const getUserTopBidsV1Options: RouteOptions = {
                           WHERE token_sets.attribute_id = attributes.id)
                       END  
                    FROM token_sets
-                   WHERE token_sets.id = y.token_set_id) 
+                   WHERE token_sets.id = y.token_set_id
+                   AND token_sets.schema_hash = y.token_set_schema_hash) 
                   ELSE NULL
                 END
               ) AS bid_context
         FROM nft_balances nb
         JOIN LATERAL (
             SELECT o.token_set_id, o.id AS "top_bid_id", o.price AS "top_bid_price", o.value AS "top_bid_value",
-                   o.maker AS "top_bid_maker", source_id_int, o.created_at "order_created_at",
+                   o.maker AS "top_bid_maker", source_id_int, o.created_at "order_created_at", o.token_set_schema_hash,
                    extract(epoch from o.created_at) * 1000000 AS "order_created_at_micro",
                    DATE_PART('epoch', LOWER(o.valid_between)) AS "top_bid_valid_from",
                    COALESCE(
@@ -264,7 +266,7 @@ export const getUserTopBidsV1Options: RouteOptions = {
             LIMIT 1
         ) y ON TRUE
         LEFT JOIN LATERAL (
-            SELECT t.token_id, t.name, t.image, t.collection_id
+            SELECT t.token_id, t.name, t.image, t.collection_id, floor_sell_value AS "token_floor_sell_value"
             FROM tokens t
             WHERE t.contract = nb.contract
             AND t.token_id = nb.token_id
@@ -312,6 +314,7 @@ export const getUserTopBidsV1Options: RouteOptions = {
             tokenId: tokenId,
             name: r.name,
             image: Assets.getLocalAssetsLink(r.image),
+            floorAskPrice: r.token_floor_sell_value ? formatEth(r.token_floor_sell_value) : null,
             collection: {
               id: r.collection_id,
               name: r.collection_name,