+++ src/api/routes.ts
@@ -562,6 +562,12 @@ export const setupRoutes = (server: Server) => {
     options: oracleEndpoints.getCollectionFloorAskOracleV2Options,
   });
 
+  server.route({
+    method: "GET",
+    path: "/oracle/collections/{collection}/floor-ask/v3",
+    options: oracleEndpoints.getCollectionFloorAskOracleV3Options,
+  });
+
   // Orders
 
   server.route({
@@ -862,6 +868,12 @@ export const setupRoutes = (server: Server) => {
     method: "GET",
     path: "/users/{user}/tokens/v4",
     options: tokensEndpoints.getUserTokensV4Options,
+  });
+
+  server.route({
+    method: "GET",
+    path: "/users/{user}/tokens/v5",
+    options: tokensEndpoints.getUserTokensV5Options,
   });
 
   server.route({